requirements.txt
feedparser
python-dateutil
requests
pyyaml
treelib